feat: 채팅 클라이언트 구현
상대 좌측, 본인 우측 표시
유저정보 표시
상단 날짜 표시

방 나가기 구현(메시지 미구현)

+++ src/main/java/com/letsparty/mapper/ChatMessageMapper.java
@@ -10,5 +10,5 @@ public interface ChatMessageMapper {
 
 	void insertChatMessage(ChatMessage chatMessage);
 	
-	void decreaseUnreadCntByRoomNoAndLastReadMessageNo(@Param("roomNo") int roomNo, @Param("lastReadMessageNo") long lastReadMessageNo);
+	void decreaseUnreadCntByRoomNoAndLastReadMessageNo(@Param("roomNo") long roomNo, @Param("lastReadMessageNo") long lastReadMessageNo);
 }
